# api-rest-app/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "API for user authentication, management, and money transfers.",
    "version": "1.0.0",
    "title": "API Rest"
  },
  "host": "localhost:3000",
  "basePath": "/api",
  "schemes": [
    "http"
  ],
  "paths": {
    "/auth/register": {
      "post": {
        "summary": "Register a new user",
        "description": "Creates a new user account.",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User registered successfully"
          },
          "400": {
            "description": "User already exists"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "User login",
        "description": "Authenticates a user and returns a token.",
        "parameters": [
          {
            "name": "credentials",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Credentials"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login successful"
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/users": {
      "get": {
        "summary": "Get all users",
        "description": "Retrieves a list of all registered users.",
        "responses": {
          "200": {
            "description": "A list of users",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/User"
              }
            }
          }
        }
      }
    },
    "/transfer": {
      "post": {
        "summary": "Transfer money",
        "description": "Processes a money transfer between users.",
        "parameters": [
          {
            "name": "transfer",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Transfer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Transfer successful"
          },
          "400": {
            "description": "Transfer failed due to validation errors"
          }
        }
      }
    }
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "favored": {
          "type": "boolean"
        }
      },
      "required": ["username", "password"]
    },
    "Credentials": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "required": ["username", "password"]
    },
    "Transfer": {
      "type": "object",
      "properties": {
        "sender": {
          "type": "string"
        },
        "recipient": {
          "type": "string"
        },
        "amount": {
          "type": "number"
        }
      },
      "required": ["sender", "recipient", "amount"]
    }
  }
}